% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benford_distribution.R
\name{benford_distribution}
\alias{benford_distribution}
\title{Distribution of Probabilities Predicted by Benford's Law}
\usage{
benford_distribution(digit = "first", base)
}
\arguments{
\item{digit}{digit from significand to use for calculating distribution of probabilities. Default is \code{"first"}
with option to also select \code{"second"}.}

\item{base}{the number base to use.}
}
\value{
a vector of probabilities corresponding to each digit.
}
\description{
Compute the distribution expected probabilities for first- or second-digit Benford's Law for any base.
}
\examples{
benford_distribution(base=10)
}
\references{
Anderson K.M., Dayaratna K., Gonshorowski D., & Miller S.J. (2022). "A New Benford Test for Clustered Data with Applications to American Elections." \emph{Stats}, 5(3), 841-855. https://doi.org/10.3390/stats5030049
}
