% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.weibull.R
\name{dd.weibull}
\alias{dd.weibull}
\title{Probability Mass Function of the Discrete Weibull Distribution}
\usage{
dd.weibull(x, a, b)
}
\arguments{
\item{x}{A vector of values corresponding to realizations of the random variable.}

\item{a}{Scale parameter of the discrete weibull distribution. See details.}

\item{b}{Shape parameter of the discrete weibull distribution. See details.}
}
\value{
A vector of values corresponding to the probability masses of the discrete weibull distribution.
}
\description{
Probability Mass Function of the Discrete Weibull Distribution
}
\details{
For this function, the parameterization for the discrete weibull distribution is:
\deqn{exp[-(\frac{x}{a})^b]-exp[-(\frac{x+1}{a})^b]}
}
\examples{
dd.weibull(1:10, a=10, b=1)
}
