% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ohio2016}
\alias{ohio2016}
\title{U.S. President Precinct-Level Returns 2016 for the State of Ohio}
\format{
\subsection{\code{ohio2016}}{

A data frame with 8,887 rows and 4 columns:
\describe{
\item{county_name}{County name}
\item{precinct}{Name of the voting precinct}
\item{clinton}{Number of votes for Hillary Clinton}
\item{trump}{Number of votes for Donald Trump}
...
}
}
}
\source{
\url{doi:10.7910/DVN/LYWX3D}
}
\usage{
ohio2016
}
\description{
Precinct-level returns for President of the United States from the 2016
General Election in the State of Ohio
}
\keyword{datasets}
