% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_digit.R
\name{extract_digit}
\alias{extract_digit}
\title{Extract a Digit from a Number}
\usage{
extract_digit(x, digit, number.of.digits = 1, last = FALSE)
}
\arguments{
\item{x}{A vector of numbers to extract digits from.}

\item{digit}{An integer corresponding to the starting digit of the numbers in x to be extracted.}

\item{number.of.digits}{Number of digits to extract. Default is 1.}

\item{last}{If \code{TRUE}, function will extract digits from left to right, i.e. starting from the last digit.
If \code{FALSE} (default), function will extract digits from right to left, i.e. starting from the first digit.}
}
\value{
An integer vector of extracted digits.
}
\description{
Extract a digit (or digits) from a vector of numbers.
}
\examples{
extract_digit(1:20, digit=1)
extract_digit(90:120, digit=2)
extract_digit(90:120, digit=1, number.of.digits=2)
extract_digit(1000:1030, number.of.digits=2, last=TRUE)
}
